+++ app/src/main/java/com/mangoboss/app/api/controller/auth/AuthController.java
@@ -1,11 +1,7 @@
 package com.mangoboss.app.api.controller.auth;
 
-import com.mangoboss.app.common.exception.CustomUserDetails;
-import com.mangoboss.app.dto.auth.requeset.SignUpRequest;
 import com.mangoboss.app.dto.auth.response.JwtResponse;
 
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
