+++ app/src/main/java/com/mangoboss/app/api/controller/contract/BossContractController.java
@@ -1,7 +1,7 @@
 package com.mangoboss.app.api.controller.contract;
 
 import com.mangoboss.app.api.facade.contract.BossContractFacade;
-import com.mangoboss.app.common.exception.CustomUserDetails;
+import com.mangoboss.app.common.security.CustomUserDetails;
 import com.mangoboss.app.dto.ListWrapperResponse;
 import com.mangoboss.app.dto.contract.request.ContractCreateRequest;
 import com.mangoboss.app.dto.contract.request.ContractTemplateCreateRequest;
