+++ app/src/main/java/com/mangoboss/app/api/controller/payroll/BossPayrollController.java
@@ -2,7 +2,7 @@ package com.mangoboss.app.api.controller.payroll;
 
 
 import com.mangoboss.app.api.facade.payroll.BossPayrollFacade;
-import com.mangoboss.app.common.exception.CustomUserDetails;
+import com.mangoboss.app.common.security.CustomUserDetails;
 import com.mangoboss.app.dto.payroll.request.AccountRegisterRequest;
 import com.mangoboss.app.dto.payroll.request.PayrollSettingRequest;
 import com.mangoboss.app.dto.payroll.response.AccountRegisterResponse;
@@ -21,14 +21,14 @@ public class BossPayrollController {
 
     @PostMapping("/account-verification")
     public AccountRegisterResponse verifyAndRegisterAccount(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                                            @PathVariable Long storeId, @RequestBody @Valid AccountRegisterRequest request){
+                                                            @PathVariable Long storeId, @RequestBody @Valid AccountRegisterRequest request) {
         final Long userId = userDetails.getUserId();
-        return bossPayrollFacade.registerBossAccount(storeId,userId,request);
+        return bossPayrollFacade.registerBossAccount(storeId, userId, request);
     }
 
     @PostMapping("/settings")
     public void updatePayrollSettings(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                            @PathVariable Long storeId, @RequestBody @Valid PayrollSettingRequest request) {
+                                      @PathVariable Long storeId, @RequestBody @Valid PayrollSettingRequest request) {
         final Long userId = userDetails.getUserId();
         bossPayrollFacade.updatePayrollSettings(storeId, userId, request);
     }
