+++ app/src/main/java/com/mangoboss/app/api/controller/schedule/BossScheduleController.java
@@ -1,18 +1,14 @@
 package com.mangoboss.app.api.controller.schedule;
 
 import com.mangoboss.app.api.facade.schedule.BossScheduleFacade;
-import com.mangoboss.app.common.exception.CustomUserDetails;
-import com.mangoboss.app.dto.ListWrapperResponse;
+import com.mangoboss.app.common.security.CustomUserDetails;
 import com.mangoboss.app.dto.schedule.request.ScheduleCreateRequest;
 import com.mangoboss.app.dto.schedule.request.ScheduleUpdateRequest;
-import com.mangoboss.app.dto.schedule.response.ScheduleDailyResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-
-import java.time.LocalDate;
 
 @RestController
 @RequiredArgsConstructor
