+++ app/src/main/java/com/mangoboss/app/api/controller/store/BossStoreController.java
@@ -10,7 +10,7 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import com.mangoboss.app.api.facade.store.BossStoreFacade;
-import com.mangoboss.app.common.exception.CustomUserDetails;
+import com.mangoboss.app.common.security.CustomUserDetails;
 import com.mangoboss.app.dto.store.request.BusinessNumberRequest;
 import com.mangoboss.app.dto.store.request.StoreCreateRequest;
